requirements.txt
streamlit
tensorflow
plotly.express
matplotlib